finished backend refactoring
