+++ seed.js
@@ -21,7 +21,6 @@ Refer to the q documentation for why and how q.invoke is used.
 
 var mongoose = require('mongoose');
 var connectToDb = require('./server/db');
-var User = mongoose.model('User');
 var q = require('q');
 var chalk = require('chalk');
